use build type specific extensions, remove duplicate code
