+++ setup.cfg
@@ -66,9 +66,15 @@ colcon_core.package_identification =
     ignore_ament_install = colcon_ros.package_identification.ignore:IgnorePackageIdentification
     ros = colcon_ros.package_identification.ros:RosPackageIdentification
 colcon_core.task.build =
-    ros = colcon_ros.task.ros:RosTask
+    ros.ament_cmake = colcon_ros.task.ament_cmake.build:AmentCmakeBuildTask
+    ros.ament_python = colcon_ros.task.ament_python.build:AmentPythonBuildTask
+    ros.catkin = colcon_ros.task.catkin.build:CatkinBuildTask
+    ros.cmake = colcon_ros.task.cmake.build:CmakeBuildTask
 colcon_core.task.test =
-    ros = colcon_ros.task.ros:RosTask
+    ros.ament_cmake = colcon_ros.task.ament_cmake.test:AmentCmakeTestTask
+    ros.ament_python = colcon_ros.task.ament_python.test:AmentPythonTestTask
+    ros.catkin = colcon_ros.task.catkin.test:CatkinTestTask
+    ros.cmake = colcon_ros.task.cmake.test:CmakeTestTask
 
 [flake8]
 import-order-style = google
